requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
ibm-cloud-sdk-core==3.22.1
ibm-platform-services==0.59.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
mpmath==1.3.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pbr==6.1.1
pillow==11.1.0
psutil==7.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.4
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
qiskit==1.4.0
qiskit-aer==0.16.1
qiskit-ibm-runtime==0.36.1
qiskit-machine-learning==0.8.2
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.16.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
six==1.17.0
stevedore==5.4.1
symengine==0.13.0
sympy==1.13.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
